--- src/Backend/Models/Response/ApplicationCreatedResponse.cs
@@ -1,5 +1,0 @@
-﻿namespace Backend.Models.Response;
-
-using System;
-
-public record ApplicationCreatedResponse(string DisplayName, Guid ClientId, string ClientSecret, string TenantID);
